stki_line_api/requirements.txt
line-bot-sdk
pip 
flask
gunicorn
psycopg2
gensim
tensorflow
keras
scikit-learn
requests
numpy
argparse